+++ cocoahttpserver/SSCrypto.m
@@ -646,7 +646,7 @@
 		unsigned long check = RSA_check_key(privateRSA);
 		if(check != 1)
 		{
-			NSLog(@"RSA_check_key() failed with result %d!", check);
+			NSLog(@"RSA_check_key() failed with result %d!", (int) check);
 			return nil;
 		}			
 		
@@ -966,7 +966,7 @@
 	unsigned long check = RSA_check_key(privateRSA);
 	if(check != 1)
 	{
-		NSLog(@"RSA_check_key() failed with result %d!", check);
+		NSLog(@"RSA_check_key() failed with result %d!", (int) check);
 		return nil;
 	}
